+++ e2e/integration/apps/forge-vite/vite.main.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from 'vite';
-import { datadogVitePlugin } from '@datadog/electron-sdk/vite-plugin';
-
-export default defineConfig({
-  plugins: [datadogVitePlugin()],
-});